refactor exports in extensions
