+++ src/extensions/components/WebhookDetailsPage/components/WebhookHeaders/WebhookHeaders.test.tsx
@@ -4,7 +4,7 @@ import { act, render, screen } from "@testing-library/react";
 import userEvent from "@testing-library/user-event";
 
 import { customHeaders } from "./utils.test";
-import WebhookHeaders, { WebhookHeadersProps } from "./WebhookHeaders";
+import { WebhookHeaders, WebhookHeadersProps } from "./WebhookHeaders";
 
 export const props: WebhookHeadersProps = {
   data: {
