+++ src/extensions/components/WebhookDetailsPage/components/WebhookInfo/WebhookInfo.tsx
@@ -12,7 +12,7 @@ import * as React from "react";
 import { FormattedMessage, useIntl } from "react-intl";
 
 import { WebhookFormData } from "../../WebhookDetailsPage";
-import WebhookStatus from "../WebhookStatus";
+import { WebhookStatus } from "../WebhookStatus/WebhookStatus";
 import { messages } from "./messages";
 import { useStyles } from "./styles";
