+++ src/extensions/hooks/useExtensions.ts
@@ -8,8 +8,8 @@ import {
 import { RelayToFlat } from "@dashboard/types";
 import { mapEdgesToItems } from "@dashboard/utils/maps";
 
-import { useExternalApp } from "../components/ExternalAppContext";
 import { AppData } from "../components/ExternalAppContext/context";
+import { useExternalApp } from "../components/ExternalAppContext/ExternalAppContext";
 import { Extension, ExtensionWithParams } from "../types";
 import { AppDetailsUrlMountQueryParams } from "../urls";
