+++ src/extensions/views/EditManifestExtension/components/AppDetailsPage/AppDetailsPage.tsx
@@ -6,7 +6,7 @@ import { ErrorBoundary } from "react-error-boundary";
 import { AppWebhooksDisplay } from "../AppWebhooksDisplay/AppWebhooksDisplay";
 import { AboutCard } from "./AboutCard";
 import { DataPrivacyCard } from "./DataPrivacyCard";
-import Header from "./Header";
+import { Header } from "./Header";
 import { PermissionsCard } from "./PermissionsCard";
 
 interface AppDetailsPageProps {
