+++ src/extensions/views/ExploreExtensions/components/ExtenionItem/hooks/useExtension.tsx
@@ -3,7 +3,7 @@ import { ExtensionData } from "@dashboard/extensions/types";
 import { Box, useTheme } from "@saleor/macaw-ui-next";
 import { useIntl } from "react-intl";
 
-import { AppActions } from "../components/AppActions";
+import { AppActions } from "../components/AppActions/AppActions";
 import { PluginActions } from "../components/PluginActions";
 
 const saleor = "Saleor Commerce";
