+++ src/extensions/views/InstallCustomExtension/components/InstallCustomExtensionFromUrl/InstallCustomExtensionFromUrl.tsx
@@ -17,7 +17,7 @@ import { useFetchManifest } from "../../hooks/useFetchManifest";
 import { useInstallApp } from "../../hooks/useInstallApp";
 import { useLoadQueryParamsToForm } from "../../hooks/useLoadQueryParamsToForm";
 import { ExtensionInstallFormData } from "../../types";
-import { InstallSectionData } from "../InstallSectionData";
+import { InstallSectionData } from "../InstallSectionData/InstallSectionData";
 import { InstallTopNav } from "../InstallTopNav";
 import { ManifestErrorMessage } from "../ManifestErrorMessage/ManifestErrorMessage";
