+++ src/extensions/views/InstallCustomExtension/components/InstallSectionData/InstallExtensionManifestData/InstallExtensionManifestData.tsx
@@ -1,5 +1,5 @@
 import { DashboardCard } from "@dashboard/components/Card";
-import { ExternalLinkUnstyled } from "@dashboard/extensions/components/ExternalLinkUnstyled";
+import { ExternalLinkUnstyled } from "@dashboard/extensions/components/ExternalLinkUnstyled/ExternalLinkUnstyled";
 import { messages } from "@dashboard/extensions/messages";
 import { ExclamationIcon } from "@dashboard/icons/ExclamationIcon";
 import { Box, Text } from "@saleor/macaw-ui-next";
