+++ src/extensions/views/ViewManifestExtension/components/AppFrame/appActionsHandler.test.ts
@@ -4,7 +4,7 @@ import { renderHook } from "@testing-library/react-hooks";
 import * as ReactIntl from "react-intl";
 import { IntlShape } from "react-intl";
 
-import * as ExternalAppContext from "../ExternalAppContext";
+import * as ExternalAppContext from "../ExternalAppContext/ExternalAppContext";
 import { AppActionsHandler } from "./appActionsHandler";
 
 jest.mock("@dashboard/config", () => {
