+++ src/index.tsx
@@ -56,7 +56,7 @@ import { CustomAppSections } from "./custom-apps/urls";
 import { CustomerSection } from "./customers";
 import DiscountSection from "./discounts";
 import { ExtensionsSection } from "./extensions";
-import { ExternalAppProvider } from "./extensions/components/ExternalAppContext";
+import { ExternalAppProvider } from "./extensions/components/ExternalAppContext/ExternalAppContext";
 import { FeatureFlagsProviderWithUser } from "./featureFlags/FeatureFlagsProvider";
 import GiftCardSection from "./giftCards";
 import { giftCardsSectionUrlName } from "./giftCards/urls";
